+++ StabilityClient.Net/StabilityClient.Net.Tests.Unit/Tests/AsyncServerStreamingCallExtensionsTests.cs
@@ -49,5 +49,43 @@ public class AsyncServerStreamingCallExtensionsTests {
         // Assert
         await saveResultsAction.Should().ThrowAsync<ArgumentException>();
     }
+    [Test]
+    public async Task SaveSingleImageToAsync_ShouldReturnResult() {
+        // Arrange
+        var fileName = "Test";
+        var response = AsyncServerStreamingCallFactory.Create(1);
+        // Act
+        var saveResult = await response.SaveSingleImageToAsync(DirectoryPath, fileName);
+        // Assert
+        
+        saveResult.FullPath.Should().Contain(Path.Combine(Path.GetFullPath(DirectoryPath), fileName))
+            .And.EndWith(".png");
+    }
+    [Test]
+    public async Task SaveSingleImageToAsync_FileNameAlreadyExist_ShouldReturnResult() {
+        // Arrange
+        var fileName = "Test";
+        var firstResponse = AsyncServerStreamingCallFactory.Create(1);
+        var firstSaveResult = await firstResponse.SaveSingleImageToAsync(DirectoryPath, fileName);
+        var secondResponse = AsyncServerStreamingCallFactory.Create(1);
 
+        // Act
+        var secondSaveResult = await secondResponse.SaveSingleImageToAsync(DirectoryPath, fileName);
+        // Assert
+        secondSaveResult.FullPath.Length.Should().BeGreaterThan(firstSaveResult.FullPath.Length);
+        secondSaveResult.FullPath.Should().Contain(Path.Combine(Path.GetFullPath(DirectoryPath), fileName))
+            .And.EndWith(".png");
+    }
+    [Test]
+    [TestCase("")]
+    [TestCase(null)]
+    public async Task SaveImageToAsync_IncorrectDirectoryPath_ShouldThrowArgumentError(string? path) {
+        // Arrange
+        var fileName = "Test";
+        var response = AsyncServerStreamingCallFactory.Create(1);
+        // Act
+        var saveResultsAction = async () => await response.SaveSingleImageToAsync(path!, fileName);
+        // Assert
+        await saveResultsAction.Should().ThrowAsync<ArgumentException>();
+    }
 }