+++ StabilityClient.Net/StabilityClient.Net.Tests.Unit/Tests/RequestBuilderTests.cs
@@ -1,6 +1,11 @@
+using Google.Protobuf;
+using Gooseai;
+
 namespace StabilityClient.Net.Tests.Unit.Tests;
 
 public class RequestBuilderTests {
+    private const string PathToTestImage = "./TestAssets/test.png";
+
     [Test]
     public void Build_WithoutSetTextPrompt_ShouldThrowArgumentException() {
         // Arrange
@@ -24,6 +29,24 @@ public class RequestBuilderTests {
         request.Prompt.Should().HaveCount(1);
         request.Prompt[0].Text.Should().Be(expectedPrompt);
         request.Prompt[0].Parameters.Init.Should().BeTrue();
+        request.Prompt[0].Parameters.Weight.Should().Be(1);
+    }
+
+    [Test]
+    public void SetTextPrompt_Weight_ShouldSetTextPrompt() {
+        // Arrange
+        var expectedPrompt = "Test";
+        var expectedWeight = 2;
+        var builder = new RequestBuilder();
+        // Act
+        var request = builder
+            .SetTextPrompt(expectedPrompt, expectedWeight)
+            .Build();
+        // Assert
+        request.Prompt.Should().HaveCount(1);
+        request.Prompt[0].Text.Should().Be(expectedPrompt);
+        request.Prompt[0].Parameters.Init.Should().BeTrue();
+        request.Prompt[0].Parameters.Weight.Should().Be(expectedWeight);
     }
 
     [Test]
@@ -41,6 +64,20 @@ public class RequestBuilderTests {
         request.Image.Width.Should().Be(RequestBuilder.DefaultImageWidth);
         request.Image.Steps.Should().Be(RequestBuilder.DefaultImageSteps);
         request.Image.Samples.Should().Be(RequestBuilder.DefaultNumberOfSamples);
+    }
+
+    [Test]
+    public void Build_SetMaskImageWithoutInitImage_ShouldThrowArgumentException() {
+        // Arrange
+        var expectedPrompt = "Test";
+        var builder = new RequestBuilder();
+        // Act
+        var act = () => builder
+            .SetTextPrompt(expectedPrompt)
+            .SetMaskImage(PathToTestImage)
+            .Build();
+        // Assert
+        act.Should().Throw<ArgumentException>();
     }
 
     [Test]
@@ -132,5 +169,138 @@ public class RequestBuilderTests {
             .Build();
         // Assert
         request.Image.Samples.Should().Be(expectedSamples);
+    }
+
+    [Test]
+    public void SetInitImage_ShouldSetInitImage() {
+        // Arrange
+        var expectedBinary = ByteString.FromStream(File.OpenRead(PathToTestImage));
+        var builder = new RequestBuilder();
+        // Act
+        var request = builder
+            .SetInitImage(PathToTestImage)
+            .Build();
+        // Assert
+        request.Prompt.Should().HaveCount(1);
+        request.Prompt[0].Artifact.Type.Should().Be(ArtifactType.ArtifactImage);
+        request.Prompt[0].Artifact.Binary.Should().BeEquivalentTo(expectedBinary);
+        request.Prompt[0].Parameters.Init.Should().BeTrue();
+        request.Prompt[0].Parameters.Weight.Should().Be(1);
+    }
+
+    [Test]
+    public void SetInitImage_Weight_ShouldSetInitImage() {
+        // Arrange
+        var expectedBinary = ByteString.FromStream(File.OpenRead(PathToTestImage));
+        var expectedWeight = 2;
+        var builder = new RequestBuilder();
+        // Act
+        var request = builder
+            .SetInitImage(PathToTestImage, expectedWeight)
+            .Build();
+        // Assert
+        request.Prompt.Should().HaveCount(1);
+        request.Prompt[0].Artifact.Type.Should().Be(ArtifactType.ArtifactImage);
+        request.Prompt[0].Artifact.Binary.Should().BeEquivalentTo(expectedBinary);
+        request.Prompt[0].Parameters.Init.Should().BeTrue();
+        request.Prompt[0].Parameters.Weight.Should().Be(expectedWeight);
+    }
+
+    [Test]
+    [TestCase("")]
+    [TestCase(null)]
+    public void SetInitImage_IncorrectPath_ShouldThrowArgumentException(string? path) {
+        // Arrange
+        var builder = new RequestBuilder();
+        // Act
+        var act = () => builder.SetInitImage(path!);
+        // Assert
+        act.Should().Throw<ArgumentException>();
+    }
+
+    [Test]
+    public void SetMaskImage_ShouldSetMaskImage() {
+        // Arrange
+        var expectedPrompt = "Test";
+        var expectedBinary = ByteString.FromStream(File.OpenRead(PathToTestImage));
+        var builder = new RequestBuilder();
+        // Act
+        var request = builder
+            .SetTextPrompt(expectedPrompt)
+            .SetInitImage(PathToTestImage)
+            .SetMaskImage(PathToTestImage)
+            .Build();
+        // Assert
+        request.Prompt.Should().HaveCount(3);
+        request.Prompt[2].Artifact.Type.Should().Be(ArtifactType.ArtifactMask);
+        request.Prompt[2].Artifact.Binary.Should().BeEquivalentTo(expectedBinary);
+    }
+
+    [Test]
+    [TestCase("")]
+    [TestCase(null)]
+    public void SetMaskImage_IncorrectPath_ShouldThrowArgumentException(string? path) {
+        // Arrange
+        var builder = new RequestBuilder();
+        // Act
+        var act = () => builder.SetMaskImage(path!);
+        // Assert
+        act.Should().Throw<ArgumentException>();
+    }
+    [Test]
+    public void SetStartSchedule_ShouldSetStartSchedule() {
+        // Arrange
+        var expectedPrompt = "Test";
+        var expectedStartSchedule = 0.3f;
+        var builder = new RequestBuilder();
+        // Act
+        var request = builder
+            .SetTextPrompt(expectedPrompt)
+            .SetStartSchedule(expectedStartSchedule)
+            .Build();
+        // Assert
+        request.Image.Parameters[0].Schedule.Start.Should().Be(expectedStartSchedule);
+    }
+    [Test]
+    public void SetEndSchedule_ShouldSetEndSchedule() {
+        // Arrange
+        var expectedPrompt = "Test";
+        var expectedEndSchedule = 0.3f;
+        var builder = new RequestBuilder();
+        // Act
+        var request = builder
+            .SetTextPrompt(expectedPrompt)
+            .SetEndSchedule(expectedEndSchedule)
+            .Build();
+        // Assert
+        request.Image.Parameters[0].Schedule.End.Should().Be(expectedEndSchedule);
+    }
+    [Test]
+    public void SetSampler_ShouldSetSampler() {
+        // Arrange
+        var expectedPrompt = "Test";
+        var expectedSampler = DiffusionSampler.SamplerKDpm2;
+        var builder = new RequestBuilder();
+        // Act
+        var request = builder
+            .SetTextPrompt(expectedPrompt)
+            .SetSampler(expectedSampler)
+            .Build();
+        // Assert
+        request.Image.Transform.Diffusion.Should().Be(expectedSampler);
+    }
+    [Test]
+    public void SetCfgScale_ShouldSetCfgScale() {
+        // Arrange
+        var expectedPrompt = "Test";
+        var expectedCfgScale = 0.3f;
+        var builder = new RequestBuilder();
+        // Act
+        var request = builder
+            .SetTextPrompt(expectedPrompt)
+            .SetCfgScale(expectedCfgScale)
+            .Build();
+        // Assert
+        request.Image.Parameters[0].Sampler.CfgScale.Should().Be(expectedCfgScale);
     }
 }