+++ StabilityClient.Net/StabilityClient.Net/Extensions/AsyncServerStreamingCallExtensions.cs
@@ -5,6 +5,9 @@ using StabilityClient.Net.Models;
 
 namespace StabilityClient.Net.Extensions;
 
+/// <summary>
+/// Usability utilities to AsyncServerStreamingCall
+/// </summary>
 public static class AsyncServerStreamingCallExtensions {
     /// <summary>
     /// Save images from response to file in expected directory.
@@ -16,11 +19,7 @@ public static class AsyncServerStreamingCallExtensions {
     /// <exception cref="ArgumentException">throw when directory path is null or empty</exception>
     public static async Task<IEnumerable<GenerateResponseSaveResult>> SaveImagesToAsync(
         this AsyncServerStreamingCall<Answer> response, string directoryPath, CancellationToken token = default) {
-        if (string.IsNullOrEmpty(directoryPath)) {
-            throw new ArgumentException(
-                $"Directory path cannot be null or empty, was: {directoryPath}. Change value of {nameof(directoryPath)}.",
-                nameof(directoryPath));
-        }
+        ValidatePath(directoryPath);
 
         var streamReader = response.ResponseStream;
         var result = new List<GenerateResponseSaveResult>();
@@ -29,6 +28,35 @@ public static class AsyncServerStreamingCallExtensions {
             var answer = streamReader.Current;
             var saveResult = await SaveArtifacts(directoryPath, answer.Artifacts, token);
             result.AddRange(saveResult);
+        }
+
+        return result;
+    }
+
+    /// <summary>
+    /// Save single image from response to file with given name in expected directory; if file with such name already exist, guid is added to the file name
+    /// </summary>
+    /// <param name="response">extended type</param>
+    /// <param name="directoryPath">path to directory where image is expected to save</param>
+    /// <param name="fileName">name for the file to saved</param>
+    /// <param name="token">cancellation token</param>
+    /// <returns>GenerateResponseSaveResult with FullPath for saved file</returns>
+    /// <exception cref="ArgumentException">throw when directory path is null or empty</exception>
+    public static async Task<GenerateResponseSaveResult> SaveSingleImageToAsync(
+        this AsyncServerStreamingCall<Answer> response, string directoryPath, string fileName,
+        CancellationToken token = default) {
+        ValidatePath(directoryPath);
+
+        var result = new GenerateResponseSaveResult();
+        var streamReader = response.ResponseStream;
+        while (await streamReader.MoveNext(token)) {
+            var answer = streamReader.Current;
+            foreach (var artifact in answer.Artifacts) {
+                if (artifact.Type == ArtifactType.ArtifactImage) {
+                    result.FullPath = await SaveImage(directoryPath, artifact.Binary.ToByteArray(), token, fileName);
+                    return result;
+                }
+            }
         }
 
         return result;
@@ -51,8 +79,9 @@ public static class AsyncServerStreamingCallExtensions {
         return result;
     }
 
-    private static async Task<string> SaveImage(string directoryPath, byte[] content, CancellationToken token) {
-        var fileName = $"{DateTime.Now:MM-dd-yy_HH-mm-ss}-{Guid.NewGuid()}.png";
+    private static async Task<string> SaveImage(string directoryPath, byte[] content, CancellationToken token,
+        string? fileName = null) {
+        fileName = GetFileName(fileName, directoryPath);
         var fullPath = Path.Combine(Path.GetFullPath(directoryPath), fileName);
 
         await File.WriteAllBytesAsync(fullPath, content, token);
@@ -63,4 +92,25 @@ public static class AsyncServerStreamingCallExtensions {
     private static GenerateResponseSaveResult CreateResponse(string path) => new() {
         FullPath = path
     };
+
+    private static string GetFileName(string? fileName, string directoryPath) {
+        if (fileName is null) {
+            return $"{DateTime.Now:MM-dd-yy_HH-mm-ss}-{Guid.NewGuid()}.png";
+        }
+
+        var fullPath = Path.Combine(Path.GetFullPath(directoryPath), $"{fileName}.png");
+        if (File.Exists(fullPath)) {
+            return $"{fileName}-{Guid.NewGuid()}.png";
+        }
+
+        return $"{fileName}.png";
+    }
+
+    private static void ValidatePath(string path) {
+        if (string.IsNullOrEmpty(path)) {
+            throw new ArgumentException(
+                $"Directory path cannot be null or empty, was: {path}. Change value of {nameof(path)}.",
+                nameof(path));
+        }
+    }
 }