+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [1.3.0] - 11.02.2023
+
+### Added
+- New parameters to RequestBuilder:
+  - init image
+  - mask image
+  - start schedule
+  - end schedule
+  - sampler
+  - CFG scale
+- SaveSingleImageToAsync: extension method to AsyncServerStreamingCall to easily save single image with given name
+
 ## [1.2.0] - 25.01.2023
 
 ### Changed